requirements.txt
PyYAML>=6.0.1
# 2.18+ has the IterableDataset.shard fix we rely on for multi-worker streams.
datasets>=2.18.0
tokenizers>=0.15.0
torch>=2.0
numpy>=1.24

